# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: metarang
    description: REST API for metarang microservices. Legacy-compatible JSON responses served via Kong → grpc-gateway → gRPC services.
    title: metarang API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: grpc-gateway direct (dev)
      url: http://localhost:8080
...
paths:
...
    /api/bank-accounts:
        get:
            operationId: get_bank-accounts
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create bank accounts
            tags:
                - KYC
        post:
            operationId: post_bank-accounts
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create bank accounts
            tags:
                - KYC
...
components:
    schemas:
        Error:
            properties:
                message:
                    type: string
            type: object
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http